Feat: Organize files in it's containing folders.
