+++ src/components/App.js
@@ -1,16 +1,24 @@
 import React from 'react';
-import './App.css';
+import Display from './Display';
+import ButtonPanel from './ButtonPanel';
+import '../styles/App.css';
 
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <p>
-          Smart calculator
-        </p>
-      </header>
-    </div>
-  );
+class App extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+
+    };
+  }
+
+  render() {
+    return (
+      <div className="App">
+        <Display />
+        <ButtonPanel />
+      </div>
+    );
+  }
 }
 
 export default App;
