+++ src/components/ButtonPanel.js
@@ -1,0 +1,20 @@
+import React from 'react';
+
+class ButtonPanel extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+
+    };
+  }
+
+  render() {
+    return (
+      <div>
+        <h1>Button Panel</h1>
+      </div>
+    );
+  }
+}
+
+export default ButtonPanel;
