+++ src/components/Display.js
@@ -1,0 +1,20 @@
+import React from 'react';
+
+class Display extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+
+    };
+  }
+
+  render() {
+    return (
+      <div className="display-container">
+        <h1>Display</h1>
+      </div>
+    );
+  }
+}
+
+export default Display;
