+++ src/styles/App.css
@@ -1,0 +1,9 @@
+.App {
+  text-align: center;
+  color: black;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+}
+
